Приглашения пользователей

+++ db/migrate/20141105215212_create_itemizations.rb
@@ -3,7 +3,7 @@ class CreateItemizations < ActiveRecord::Migration
     create_table :itemizations do |t|
       t.integer :item_id, index: true
       t.integer :parent_id, index: true
-      t.integer :quantity
+      t.float :quantity
       t.integer :account_id, null: false, index: true
 
       t.timestamps null: false
